Add tests

+++ app/titanic.py
@@ -6,7 +6,7 @@ class TitanicSurvival(Resource):
     def __init__(self):
         self.parser = reqparse.RequestParser()
         self.parser.add_argument('sex', type=str, required=True)
-        self.parser.add_argument('age', type=int, required=False)
+        self.parser.add_argument('age', type=float, required=False)
         self.parser.add_argument('sib_sp', type=int, required=True)
         self.parser.add_argument('ticket_class', type=int, required=True)
         self.parser.add_argument('embarked', type=str, required=True)
